Prompt user to confirm deletion of a recipe

+++ public/scripts/controllers/recipes.js
@@ -24,6 +24,11 @@
 			};
 
 			$scope.deleteRecipe = function (recipe) {
+				const confirmed = confirm(`Are you sure you want to delete recipe "${recipe.name}"?`);
+				if (!confirmed) {
+					return;
+				}
+
 				DataService.removeRecipe(recipe._id, error => {
 					if (error) {
 						return console.log(error);
